+++ docs/conf.py
@@ -15,13 +15,33 @@ release = '0.0.0a3'
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = [
-    'sphinx_copybutton',
-    'sphinx_issues'
+	'sphinx.ext.autodoc',
+	'sphinx.ext.autosummary',
+	'sphinx.ext.napoleon',
+	'sphinx.ext.viewcode',
+	'sphinx_copybutton',
+	'sphinx_issues',
 ]
 
 templates_path = ['_templates']
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
+# -- Napoleon settings -------------------------------------------------------
+napoleon_google_docstring = True
+napoleon_numpy_docstring = True
+napoleon_include_init_with_doc = False
+napoleon_include_private_with_doc = False
+
+# -- Autodoc settings --------------------------------------------------------
+autodoc_member_order = 'bysource'
+autodoc_default_flags = ['members']
+autosummary_generate = True
+
+# -- Intersphinx mapping -----------------------------------------------------
+intersphinx_mapping = {
+	'python': ('https://docs.python.org/3', None),
+	'lxml': ('https://lxml.de/', None),
+}
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -30,7 +50,12 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 html_theme = 'furo'
 html_static_path = ['_static']
 
+# Add source link in footer
+html_show_sourcelink = True
+html_copy_source = True
+html_show_sphinx = True
+
 # -- Linking Github issues --------------------------------------------------
 # https://github.com/sloria/sphinx-issues
 
-issues_github_path = "ernestofgonzalez/epub-utils"
+issues_github_path = 'ernestofgonzalez/epub-utils'