+++ epub_utils/__main__.py
@@ -1,4 +1,4 @@
 from epub_utils.cli import main
 
-if __name__ == "__main__":
-    main(prog_name="epub-utils")
+if __name__ == '__main__':
+	main(prog_name='epub-utils')