+++ docs/changelog.rst
@@ -4,23 +4,43 @@
 Changelog
 =========
 
+.. _v_0_0_0a5:
+
+0.0.0a5 (2025-06-01)
+--------------------
+
+* Added file retrieval by file path. (:issue:`22`)
+* Added pretty printing to XML inspection (:issue:`23`)
+
+.. _v_0_0_0a4:
+
+0.0.0a4 (2025-05-26)
+--------------------
+
+* Added file inspection and ``files`` CLI command. (`#20 <https://github.com/ernestofgonzalez/epub-utils/pull/20>`__)
+* Added content inspection and ``content`` CLI command (:issue:`5`)
+* Added manifest parsing and ``manifest`` CLI command (`#13 <https://github.com/ernestofgonzalez/epub-utils/pull/13>`__)
+* Added spine parsing and ``spine`` CLI command (`#9 <https://github.com/ernestofgonzalez/epub-utils/pull/9>`__)
+* Added Key-value support for ``metadata`` CLI command 
+* Fixed table of contents parsing for OEBPS 1 (`#11 <https://github.com/ernestofgonzalez/epub-utils/pull/11>`__). Thanks, `Christian Klein <https://github.com/cklein>`__.
+
 .. _v_0_0_0a3:
 
 0.0.0a3 (2025-05-04)
-------------------
+--------------------
 
 * Fixed `toc` command. (:issue:`1`)
 
 .. _v_0_0_0a2:
 
 0.0.0a2 (2025-05-03)
-------------------
+--------------------
 
 * Added classifiers
 
 .. _v_0_0_0a1:
 
 0.0.0a1 (2025-05-03)
-------------------
+--------------------
 
 * Initial relese to PyPI